pyproject.toml
[project]
name = "coursework-habits"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.8.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "telebot (>=0.0.5,<0.0.6)",
    "redis (>=6.4.0,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^7.0.0"

